Move most math stuff from support.h into math_utils.h
This is part of larger piece of work of getting rid of "kitchen sink"
style utility/helper files.
