+++ include/support.h
@@ -109,43 +109,6 @@ T to_finite(const std::string& input) {
 // Works with both \ and / directory delimiters.
 std::string get_basename(const std::string& filename);
 
-// Unsigned-only integer division with ceiling
-template<typename T1, typename T2>
-inline constexpr T1 ceil_udivide(const T1 x, const T2 y) noexcept {
-	static_assert(std::is_unsigned<T1>::value, "First parameter should be unsigned");
-	static_assert(std::is_unsigned<T2>::value, "Second parameter should be unsigned");
-	return (x != 0) ? 1 + ((x - 1) / y) : 0;
-	// https://stackoverflow.com/a/2745086
-}
-
-// Signed-only integer division with ceiling
-template<typename T1, typename T2>
-inline constexpr T1 ceil_sdivide(const T1 x, const T2 y) noexcept {
-	static_assert(std::is_signed<T1>::value, "First parameter should be signed");
-	static_assert(std::is_signed<T2>::value, "Second parameter should be signed.");
-	return x / y + (((x < 0) ^ (y > 0)) && (x % y));
-	// https://stackoverflow.com/a/33790603
-}
-
-template <typename T>
-std::function<T()> CreateRandomizer(const T min_value, const T max_value);
-
-inline int iround(double x)
-{
-	assert(std::isfinite(x));
-	assert(x >= (std::numeric_limits<int>::min)());
-	assert(x <= (std::numeric_limits<int>::max)());
-	return static_cast<int>(round(x));
-}
-
-inline int iroundf(const float x)
-{
-	assert(std::isfinite(x));
-	assert(x >= static_cast<float>((std::numeric_limits<int>::min)()));
-	assert(x <= static_cast<float>((std::numeric_limits<int>::max)()));
-	return static_cast<int>(roundf(x));
-}
-
 // Select the nearest unsigned integer capable of holding the given bits
 template <int n_bits>
 using nearest_uint_t = \
@@ -168,46 +131,6 @@ using next_uint_t = typename std::conditional<
         sizeof(T) == sizeof(uint8_t),
         uint16_t,
         typename std::conditional<sizeof(T) == sizeof(uint16_t), uint32_t, uint64_t>::type>::type;
-
-// Left-shifts a signed value by a given amount, with overflow detection
-template <typename T1, typename T2>
-constexpr T1 left_shift_signed(T1 value, T2 amount)
-{
-	// Ensure we're using a signed type
-	static_assert(std::is_signed<T1>::value, "T1 must be signed");
-
-	// Ensure the two types are integers
-	static_assert(std::numeric_limits<T1>::is_integer, "T1 must be an integer type");
-	static_assert(std::numeric_limits<T2>::is_integer, "T2 must be an integer type");
-
-	// Ensure the amount we're shifting isn't negative
-	assert(amount >= 0);
-
-	// Ensure the amount we're shifting doesn't exceed the value's bit-size
-	assert(amount <= std::numeric_limits<T1>::digits);
-
-#if defined(NDEBUG)
-	// For release builds, simply cast the value to the unsigned-equivalent
-	// to ensure performance isn't impacted. Debug and UBSAN builds catch issues.
-	typedef typename std::make_unsigned<T1>::type unsigned_T1;
-	const auto shifted = static_cast<unsigned_T1>(value) << amount;
-
-#else
-	// Ensure we can accommodate the value being shifted
-	static_assert(sizeof(T1) <= sizeof(next_uint_t<T1>),
-	              "T1 cannot be a larger than its next larger type");
-
-	// cast the value to the next larger unsigned-type before shifting
-	const auto shifted = static_cast<next_uint_t<T1>>(value) << amount;
-
-	// Ensure the value is in-bounds of its signed limits
-	assert(static_cast<next_int_t<T1>>(shifted) >= (std::numeric_limits<T1>::min)());
-	assert(static_cast<next_int_t<T1>>(shifted) <= (std::numeric_limits<T1>::max)());
-#endif
-
-	// Cast it back to its original type
-	return static_cast<T1>(shifted);
-}
 
 template <typename cast_t, typename check_t>
 constexpr cast_t check_cast(const check_t in)
@@ -232,6 +155,9 @@ constexpr cast_t check_cast(const check_t in)
 	return static_cast<cast_t>(in);
 }
 
+template <typename T>
+std::function<T()> CreateRandomizer(const T min_value, const T max_value);
+
 // Include a message in assert, similar to static_assert:
 #define assertm(exp, msg) assert(((void)msg, exp))
 // Use (void) to silent unused warnings.
@@ -243,14 +169,6 @@ constexpr cast_t check_cast(const check_t in)
 #ifdef HAVE_STRINGS_H
 #include <strings.h>
 #endif
-
-// Clamp: given a value that can be compared with the given minimum and maximum
-//        values, this function will:
-//          * return the value if it's in-between or equal to either bounds, or
-//          * return either bound depending on which bound the value is beyond
-template <class T> T clamp(const T& n, const T& lower, const T& upper) {
-	return std::max<T>(lower, std::min<T>(n, upper));
-}
 
 // Scans the provided command-line string for the '/'flag and removes it from
 // the string, returning if the flag was found and removed.
@@ -288,22 +206,6 @@ constexpr uint8_t DOS_DATE_months[] = {0,  31, 28, 31, 30, 31, 30,
 bool is_date_valid(const uint32_t year, const uint32_t month, const uint32_t day);
 
 bool is_time_valid(const uint32_t hour, const uint32_t minute, const uint32_t second);
-
-/*
-  Returns a number wrapped between the lower and upper bounds.
-   - wrap(-1, 0, 4); // Returns 4
-   - wrap(5, 0, 4); // Returns 0
-
-  All credit to Charles Bailey, https://stackoverflow.com/a/707426
-*/
-constexpr int wrap(int val, int const lower_bound, int const upper_bound)
-{
-	const auto range_size = upper_bound - lower_bound + 1;
-	if (val < lower_bound)
-		val += range_size * ((lower_bound - val) / range_size + 1);
-
-	return lower_bound + (val - lower_bound) % range_size;
-}
 
 struct FILE_closer {
 	void operator()(FILE *f) noexcept;
@@ -375,43 +277,5 @@ constexpr auto enum_val(enum_t e)
 	static_assert(std::is_enum_v<enum_t>);
 	return static_cast<std::underlying_type_t<enum_t>>(e);
 }
-
-double decibel_to_gain(const double decibel);
-double gain_to_decibel(const double gain);
-
-template <typename T>
-constexpr T lerp(const T a, const T b, const T t)
-{
-	return a * (1 - t) + b * t;
-}
-
-template <typename T>
-constexpr T invlerp(const T a, const T b, const T v)
-{
-	return (v - a) / (b - a);
-}
-
-template <typename T>
-constexpr T remap(const T in_min, const T in_max, const T out_min,
-                  const T out_max, const T v)
-{
-	const auto t = invlerp(in_min, in_max, v);
-	return lerp(out_min, out_max, t);
-}
-
-// Explicit instantiations for lerp, invlerp, and remap
-
-template float lerp<float>(const float a, const float b, const float t);
-template double lerp<double>(const double a, const double b, const double t);
-
-template float invlerp<float>(const float a, const float b, const float v);
-template double invlerp<double>(const double a, const double b, const double v);
-
-template float remap<float>(const float in_min, const float in_max,
-                            const float out_min, const float out_max, const float v);
-
-template double remap<double>(const double in_min, const double in_max,
-                              const double out_min, const double out_max,
-                              const double v);
 
 #endif
