+++ src/dos/cdrom_image.cpp
@@ -45,7 +45,7 @@
 #include "fs_utils.h"
 #include "setup.h"
 #include "string_utils.h"
-#include "support.h"
+#include "math_utils.h"
 
 using namespace std;
