+++ src/gui/sdlmain.cpp
@@ -57,6 +57,7 @@
 #include "joystick.h"
 #include "keyboard.h"
 #include "mapper.h"
+#include "math_utils.h"
 #include "mixer.h"
 #include "mouse.h"
 #include "pacer.h"
@@ -65,7 +66,6 @@
 #include "sdlmain.h"
 #include "setup.h"
 #include "string_utils.h"
-#include "support.h"
 #include "timer.h"
 #include "tracy.h"
 #include "vga.h"
