+++ src/hardware/mpu401.cpp
@@ -20,12 +20,12 @@
 
 #include <string.h>
 
+#include "cpu.h"
 #include "inout.h"
+#include "math_utils.h"
+#include "midi.h"
 #include "pic.h"
 #include "setup.h"
-#include "cpu.h"
-#include "support.h"
-#include "midi.h"
 
 static void MPU401_Event(uint32_t);
 static void MPU401_Reset();
