+++ src/hardware/pcspeaker_impulse.cpp
@@ -26,6 +26,7 @@
 #include "pcspeaker_impulse.h"
 
 #include "checks.h"
+#include "math_utils.h"
 
 CHECK_NARROWING();
