+++ tests/vs/tests.vcxproj
@@ -333,6 +333,9 @@ $(TargetPath)</Command>
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\..\src\libs\ghc\fs_std_impl.cpp" />
+    <ClCompile Include="..\..\src\libs\loguru\loguru.cpp" />
+    <ClCompile Include="..\..\src\libs\whereami\whereami.c" />
     <ClCompile Include="..\..\src\misc\ansi_code_markup.cpp" />
     <ClCompile Include="..\..\src\misc\cross.cpp" />
     <ClCompile Include="..\..\src\misc\fs_utils_win32.cpp" />
@@ -340,14 +343,12 @@ $(TargetPath)</Command>
     <ClCompile Include="..\..\src\misc\setup.cpp" />
     <ClCompile Include="..\..\src\misc\string_utils.cpp" />
     <ClCompile Include="..\..\src\misc\support.cpp" />
-    <ClCompile Include="..\..\src\libs\ghc\fs_std_impl.cpp" />
-    <ClCompile Include="..\..\src\libs\loguru\loguru.cpp" />
-    <ClCompile Include="..\..\src\libs\whereami\whereami.c" />
     <ClCompile Include="..\ansi_code_markup_tests.cpp" />
-    <ClCompile Include="..\bitops_tests.cpp" />
     <ClCompile Include="..\bit_view_tests.cpp" />
+    <ClCompile Include="..\bitops_tests.cpp" />
     <ClCompile Include="..\fs_utils_tests.cpp" />
     <ClCompile Include="..\iohandler_containers_tests.cpp" />
+    <ClCompile Include="..\math_utils_tests.cpp" />
     <ClCompile Include="..\rwqueue_tests.cpp" />
     <ClCompile Include="..\setup_tests.cpp" />
     <ClCompile Include="..\string_utils_tests.cpp" />
